Resolve battery terminal value in component
